rename and update config orders (nothing changed!)
