+++ pgdrive/world/top_down_observation/top_down_observation.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import gym
 import numpy as np
 from pgdrive.constants import Decoration, DEFAULT_AGENT, PG_EDITION
-from pgdrive.rl.observation_type import ObservationType
+from pgdrive.rl_utils.observation_type import ObservationType
 from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 from pgdrive.utils import import_pygame
 from pgdrive.world.top_down_observation.top_down_obs_impl import WorldSurface, ObservationWindow, COLOR_BLACK, \
